Better handling of REY language of administration
If the language of administration is removed from the list of languages
the test-entry is associated with then the test entry's language is
automatically changed.

Also, if all variants are of a different language to the language of
administration (and there are at least 2 variants selected) then pop up
a warning to the user.