Fixing bug that sometimes prevents supervisors from setting word details
This was caused because, once a word's misspelled, aft and fas
properties are set the word becomes locked.  However, as soon as the
model's values are set the onPatch function cannot be called, causing
the word's last empty property from actually being set on the server
side.

The solution is to create a new property, wordLocked, which is only set
AFTER the record is changed on the server side.